修复下单后再取消无法再下单的问题，并且优化搜索性能

+++ src/main/java/me/ddggdd135/slimeae/api/items/MEStorageCellCache.java
@@ -143,7 +143,7 @@ public class MEStorageCellCache implements IStorage {
         stored += toAdd;
         storageData.setStored(stored);
         storages.putKey(key, amount + toAdd);
-        itemInfo.setAmount((int) (itemInfo.getAmount() - toAdd));
+        itemInfo.setAmount(itemInfo.getAmount() - toAdd);
         SlimeAEPlugin.getStorageCellStorageDataController().markDirty(storageData);
         trim(key);
     }
